Commit41

+++ lib/modules/list_get_other_account_deposit/domain/entities/list_get_other_account_deposit_entity.dart
@@ -17,7 +17,7 @@ class ListGetOtherDccountDepositEntity {
   final int idcState;
   final String state;
   final String codMoney;
-  final double monto;
+  final int monto;
   final String thirdAccount;
   final String thirdName;
   final String accountFunds;
